Changes related to pull request review, MiddlewareConfigurationManager moved to own file, removed MiddlewareHostFilteringOptions object.
